move e2e api out of sveltekit
