+++ e2e/_api/package.json
@@ -1,0 +1,18 @@
+{
+	"name": "e2e-api",
+	"private": true,
+	"version": "1.0.0",
+	"description": "",
+	"bin": "./server.mjs",
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"graphql": "^15.5.0",
+		"@kitql/helper": "^0.5.0",
+		"graphql-relay": "^0.10.0",
+		"graphql-ws": "^5.8.2",
+		"@graphql-yoga/node": "^2.13.13",
+		"ws": "^8.8.1"
+	}
+}
