+++ e2e/_api/schema-hello.graphql
@@ -1,3 +1,3 @@
 extend type Query {
-  hello: String!
+	hello: String!
 }
