+++ e2e/react/houdini.config.js
@@ -1,1 +1,3 @@
-export default {}
+export default {
+	schemaPath: '../api/*.graphql',
+}
