+++ e2e/sveltekit/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "web": "cross-env TZ=utc vite dev --host 0.0.0.0",
-    "api": "cross-env TZ=utc node api/server.mjs",
+    "api": "cross-env e2e-api",
     "dev": "concurrently \"pnpm run web\" \"pnpm run api\" -n \"web,api\" -c \"green,magenta\"",
     "build": "vite build",
     "package": "svelte-kit package",
@@ -27,10 +27,10 @@
     "@typescript-eslint/parser": "^5.10.1",
     "concurrently": "7.1.0",
     "cross-env": "^7.0.3",
+    "e2e-api": "workspace:^",
     "eslint": "^8.12.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-svelte3": "^4.0.0",
-    "graphql": "^15.5.0",
     "houdini": "workspace:^",
     "houdini-svelte": "workspace:^",
     "prettier": "^2.5.1",
@@ -41,14 +41,5 @@
     "tslib": "^2.3.1",
     "typescript": "~4.6.2",
     "vite": "^3.1.0"
-  },
-  "dependencies": {
-    "@graphql-yoga/node": "^2.13.13",
-    "@kitql/helper": "^0.5.0",
-    "graphql-relay": "^0.10.0",
-    "graphql-tag": "^2.12.6",
-    "graphql-ws": "^5.8.2",
-    "mdsvex": "^0.10.5",
-    "ws": "^8.8.1"
   }
 }
